requirements.txt
fastapi==0.101.0
azure-core==1.28.0
botbuilder-core==4.14.4
azure-ai-language-questionanswering==1.1.0
mysql-connector-python==8.1.0
aiohttp==3.8.5
openai==0.27.8
cookiecutter==1.7.0
langchain==0.0.240
chromadb==0.4.5
selenium==4.11.2
unstructured==0.9.2
tiktoken==0.4.0
